export mixins, combineContainerMixins, and createContainer; rm VictoryVoronoiZoomContainer; demo createContainer
