+++ demo/index.html
@@ -38,9 +38,8 @@
     <p class="browsehappy">You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.</p>
   <![endif]-->
   <div id="content">
-    <h1>If you can see this, something is broken (or JS is not enabled)!!.</h1>
+    <h3>Loading...</h3>
   </div>
   <script type="text/javascript" src="assets/main.js"></script>
 </body>
 </html>
-
