--- src/components/containers/victory-voronoi-zoom-container.js
@@ -1,3 +1,0 @@
-import createContainer from "./create-container";
-
-export default createContainer("zoom", "voronoi");
